Add and install Joi as a dependency for validation of user input
