+++ package.json
@@ -17,7 +17,8 @@
   },
   "homepage": "https://github.com/CodenNerd/FastFoodFast#readme",
   "dependencies": {
-    "express": "^4.16.3"
+    "express": "^4.16.3",
+    "joi": "^13.6.0"
   },
   "devDependencies": {
     "chai": "^4.1.2",
