+++ router/api1.js
@@ -1,5 +1,6 @@
 const express = require('express');
 const orders = require('../data/orders');
+const newOrder = require('../data/newOrder');
 
 const api = express.Router();
 api.use(express.json());
@@ -9,23 +10,13 @@ api.get('/orders', (req, res) => {
 });
 
 api.get('/orders/:id', (req, res) => {
-  const order = orders.find(o => o.orderId === Number(req.params.id));
-  if (!order) res.status(404).send('That particular order was not found on our server');
+  const thisOrder = orders.find(o => o.orderId === Number(req.params.id));
+  if (!thisOrder) res.status(404).send('That particular order was not found on our server');
 
-  res.status(404).send(order);
+  res.status(404).send(thisOrder);
 });
 
-api.post('/orders', (req, res) => {
-  const d = new Date();
-  const order = {
-    orderId: orders.length + 1,
-    userId: req.body.userId,
-    food: req.body.food,
-    status: 'pending',
-    date: `${d.getDay()}/${d.getMonth()}/${d.getFullYear()}`,
-    time: `${d.getHours()}:${d.getMinutes()}`,
-  };
-  orders.push(order);
-  res.status(200).send(orders);
-});
+api.post('/orders', newOrder);
+
+
 module.exports = api;
